Move around the code in ForwardDiff a bit
