+++ ext/LuxForwardDiffExt/LuxForwardDiffExt.jl
@@ -1,0 +1,20 @@
+module LuxForwardDiffExt
+
+using ADTypes: AutoForwardDiff
+using ChainRulesCore: ChainRulesCore
+using Lux: Lux
+using FastClosures: @closure
+using ForwardDiff: ForwardDiff
+using Functors: fmap
+
+const CRC = ChainRulesCore
+
+@inline Lux._is_extension_loaded(::Val{:ForwardDiff}) = true
+
+include("utils.jl")
+
+include("jvp.jl")
+include("nested_ad.jl")
+include("batched_ad.jl")
+
+end
